move files and rename namespace.
