+++ osu.ElasticIndexer.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ElasticIndex", "ElasticIndex\ElasticIndex.csproj", "{AEE6C586-C534-47EF-9557-FE77B772D6E4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu.ElasticIndexer", "osu\ElasticIndexer\osu.ElasticIndexer.csproj", "{AEE6C586-C534-47EF-9557-FE77B772D6E4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
