+++ osu/ElasticIndexer/appsettings.json
