+++ osu/ElasticIndexer/schemas/high_scores.json
